Update patch set 1
Patch Set 1:

allthough i am not a very big fan of that shared init. i always prefer init to be device specific and have very strict control what props exactly are set for that device. it also makes sharing of devices trees easier because of less dependencies. but since that repo is optional and devices must use omni.dependencies to use it i can live with it ☺

Patch-set: 1