Update patch set 1
Patch Set 1:

Well I only uploaded the files I needed, but I can merge in all the latest CAF stuff (it seems to build fine). And this is really for devices that share everything but props, i.e. devices with multiple variants like a lot of the HTC devices. Seems like it makes more sense to have a single common repo and use the Unified flag and custom init than to have a bunch of separate repos just for a few props.

Patch-set: 1