Update patch set 2
Patch Set 2:

> > Which leads to the real question - why someone actually needs to
 > > use this libinit?
 > 
 > Mostly for device unification. In CM I use it for Moto G phones as
 > we have unified tree for both single and dual sim models.

well - I also use libinit but since that is always very device specific it makes a lot of sense to have it only in the device tree instead of here. That makes porting a device a hell lot of easier too cause it contains all it needs. Unification is not always better :)

Patch-set: 2